refactor(docker-tests): extract UTXO ordermatching tests to dedicated module
Extract 17 UTXO-only ordermatching tests from docker_tests_inner.rs to
new utxo_ordermatch_v1_tests.rs module, gated by docker-tests-ordermatch
feature flag.

Extracted tests:
- Order lifecycle (balance-driven cancellations/updates)
- Partial fill handling
- Order volume (set_price_max)
- Restart/persistence (kick_start tests)
- Same private key edge cases
- Order conversion (taker to maker)
- Best price matching (buy/sell)
- RPC response format validation

This separation allows UTXO ordermatching tests to run independently
from ETH tests, reducing CI job coupling and improving test isolation.
